[STORM-307] Implement st2 run in CLI
Add "st2 run" to perform exact function as "st2 action execute".

Closes: STORM-307
