+++ st2client/st2client/models/action.py
@@ -1,22 +1,22 @@
 import logging
 
-from st2client import models
+from st2client.models import core
 
 
 LOG = logging.getLogger(__name__)
 
 
-class ActionType(models.Resource):
+class ActionType(core.Resource):
     _display_name = 'Action Type'
     _plural = 'ActionTypes'
     _plural_display_name = 'Action Types'
 
 
-class Action(models.Resource):
+class Action(core.Resource):
     _plural = 'Actions'
 
 
-class ActionExecution(models.Resource):
+class ActionExecution(core.Resource):
     _alias = 'Execution'
     _display_name = 'Action Execution'
     _plural = 'ActionExecutions'
