+++ st2client/st2client/models/reactor.py
@@ -1,14 +1,14 @@
 import logging
 
-from st2client import models
+from st2client.models import core
 
 
 LOG = logging.getLogger(__name__)
 
 
-class Trigger(models.Resource):
+class Trigger(core.Resource):
     _plural = 'Triggers'
 
 
-class Rule(models.Resource):
+class Rule(core.Resource):
     _plural = 'Rules'
